WIP - tests not all green yet but functional
