+++ inlang/source-code/sdk/multi-project-test/project4-dir/project.inlang/messages.json
@@ -5,6 +5,16 @@
     "selectors": [],
     "variants": [
       {
+        "languageTag": "de",
+        "match": [],
+        "pattern": [
+          {
+            "type": "Text",
+            "value": "Mock translate local en to de: Generated message (1)"
+          }
+        ]
+      },
+      {
         "languageTag": "en",
         "match": [],
         "pattern": [
@@ -22,6 +32,16 @@
     "selectors": [],
     "variants": [
       {
+        "languageTag": "de",
+        "match": [],
+        "pattern": [
+          {
+            "type": "Text",
+            "value": "Mock translate local en to de: Generated message (2)"
+          }
+        ]
+      },
+      {
         "languageTag": "en",
         "match": [],
         "pattern": [
@@ -38,6 +58,16 @@
     "id": "project4_message_key_3",
     "selectors": [],
     "variants": [
+      {
+        "languageTag": "de",
+        "match": [],
+        "pattern": [
+          {
+            "type": "Text",
+            "value": "Mock translate local en to de: Generated message (3)"
+          }
+        ]
+      },
       {
         "languageTag": "en",
         "match": [],
